zh-ebn-report/pyproject.toml
[project]
name = "zh-ebn-report"
version = "0.1.0"
description = "End-to-end HITL pipeline for Taiwan evidence-based nursing (EBN) reports: vague topic → refined PICO → multi-database search → CASP appraisal → Quarto DOCX draft."
readme = "SKILL.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "htlin" },
]
keywords = [
    "evidence-based-nursing",
    "taiwan",
    "nursing-report",
    "literature-review",
    "pico",
    "casp",
    "clinical-appraisal",
]
dependencies = [
    "anthropic>=0.40.0",
    "typer>=0.12.0",
    "pydantic>=2.9.0",
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",
    "python-dotenv>=1.0.0",
    "python-frontmatter>=1.1.0",
    "bibtexparser>=1.4.0",
    "rich>=13.7.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[project.scripts]
zh-ebn-report = "zh_ebn_report.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/zh_ebn_report"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "PL"]
ignore = ["PLR0913"]  # too many args is OK for pipeline functions

[tool.mypy]
strict = true
python_version = "3.11"
plugins = ["pydantic.mypy"]
